Rename icon because they all are components
