+++ client/src/components/icons/DrinksIcon.js
