+++ client/src/components/icons/FoodIcon.js
