+++ client/src/components/icons/HotelIcon.js
