+++ client/src/components/icons/MapIcon.js
