+++ client/src/components/icons/WaveIcon.js
@@ -1,0 +1,44 @@
+import React from 'react';
+import styled from '@emotion/styled';
+
+const Svg = styled.svg``;
+
+export default function WaveIcon(props) {
+  return (
+    <Svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="375"
+      height="52.227"
+      viewBox="0 0 375 52.227"
+      {...props}
+    >
+      <g id="Gruppe_30" data-name="Gruppe 30" transform="translate(-1739.944 -1536.791)">
+        <g id="Gruppe_29" data-name="Gruppe 29" transform="translate(1740 1543.159)">
+          <path
+            id="Pfad_112"
+            data-name="Pfad 112"
+            d="M0,2.507s40.258-18.66,133.956-18.66C197.82-16.115,229.974-.8,284.787,2.507s90,0,90,0l.152,27.1-374.882.1Z"
+            transform="translate(0 16.153)"
+            fill="rgba(255,255,255,0.4)"
+          />
+        </g>
+        <g id="Gruppe_28" data-name="Gruppe 28" transform="translate(1739.944 1536.791)">
+          <path
+            id="Pfad_114"
+            data-name="Pfad 114"
+            d="M.01-85.8s200.44,40.062,271.809,40.406,103.069-8.1,103.069-8.1v19.784l-374.944.13Z"
+            transform="translate(0.056 85.8)"
+            fill="rgba(255,255,255,0.4)"
+          />
+        </g>
+        <path
+          id="Pfad_139"
+          data-name="Pfad 139"
+          d="M-.086,28.2S178.762-23.983,271.051-23.983,369.07-12.82,369.07-12.82V28.2Z"
+          transform="translate(1745.786 1560.801)"
+          fill="rgba(255,255,255,0.4)"
+        />
+      </g>
+    </Svg>
+  );
+}
