--- client/src/components/icons/seemore.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import styled from '@emotion/styled';
-import light from '../../themes/light.js';
-
-const Svg = styled.svg`
-  fill: ${light.colors.action};
-  width: 25px;
-  height: 25px;
-  margin-bottom: 45px;
-`;
-
-export default function SeeMoreIcon(props) {
-  return (
-    <Svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" {...props}>
-      <path d="M0 0h24v24H0z" fill="none" />
-      <path d="M6 10c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2zm12 0c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2zm-6 0c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2z" />
-    </Svg>
-  );
-}
