--- client/src/components/icons/wave.js
@@ -1,42 +1,0 @@
-import React from 'react';
-import styled from '@emotion/styled';
-
-const Svg = styled.svg``;
-
-export default function WaveIcon(props) {
-  return (
-    <Svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="375"
-      height="61.785"
-      viewBox="0 0 375 61.785"
-      {...props}
-    >
-      <g id="Gruppe_18" data-name="Gruppe 18" transform="translate(-1785.944 -1963.791)">
-        <g id="Gruppe_17" data-name="Gruppe 17" transform="translate(1786 1814)">
-          <path
-            id="Pfad_112"
-            data-name="Pfad 112"
-            d="M0,6.182S40.258-16.153,133.956-16.153c63.864.046,96.018,18.382,150.832,22.335s90,0,90,0l.152,32.439L.062,38.737Z"
-            transform="translate(0 172.312)"
-            fill="rgba(255,255,255,0.4)"
-          />
-          <path
-            id="Pfad_113"
-            data-name="Pfad 113"
-            d="M0,15.3s181.7-55.1,271.756-55.693,103.113,5.035,103.113,5.035v50L0,14.612Z"
-            transform="translate(0 196.272)"
-            fill="rgba(255,255,255,0.4)"
-          />
-          <path
-            id="Pfad_114"
-            data-name="Pfad 114"
-            d="M.01-85.8s200.4,46.99,271.755,47.394,103.049-9.5,103.049-9.5V-24.7l-374.87.153Z"
-            transform="translate(0 235.591)"
-            fill="rgba(255,255,255,0.4)"
-          />
-        </g>
-      </g>
-    </Svg>
-  );
-}
